Create generic integration test
